Large Network Bug Resolution; painting optimization
Fixed the large network bug! It actually was a data problem as
previously thought, but I had also done a bunch of things in
NetworkExplorer for debugging so this commit probably has a few changes
from previous, as well as a deletion of a bunch of debugging checks.

Areas of the network now only repaint when necessary: a node needs
repainting if its selection state, passability, and/or accessibility
changes. The recursive repainting function returns if it reaches a node
that is fully unselected (not selected, not on path to root) that does
not need repaint. Otherwise, the next repaint is called because it is
possible that children require a repaint. However, if a node does not
need repaint, its nodeDrawing and outgoing stream segments will not be
repainted.