Selection Bug; pan, zoom tweaks
Added colors during selection and deselection to help debug  the
inability to select different parts of the network. Found that the
problem is in the original data: some parts of the network are
reachable from multiple paths from the root. This causes the entire
subnetwork rooted at those nodes to be un-selectable. When marking the
path to root and deselecting everything else, the should-be-selected
network will always be treated as an offshoot network that needs to be
deselected.

Workaround: mark the path to the root/deselect offshoots, and then
select the subnetwork. This results in a broken path to the root: part
of it is deselected.

Added field to data file: displayInfo. Right now, display info gives
the largest x,y value so that the website can set its visual display
more intelligently.