Selection Bug Fix
Fixed selection and deselection to compensate for nodes that are
reachable through multiple paths, though the path to root marking will
only ever mark one path (the path through the one stored parent).

I have commented out the colors I was setting for debugging. If you
want to see the colors, search for "//*tmpDebugSelection". Selected
nodes ares yellow; path to root nodes are green; deselected nodes are
purple. These colors go away once the network is repainted, but in some
cases the purple colors remain: it is possible that unselected nodes
are not in need of a repaint.