Display Changes; Pan improvement
Dams: orange for no action, purple for action.
Crossings: yellow for no action, green for action.
Radius of dam is larger than or crossing, and radius for any barrier
increases if an action is taken.
Streams display slightly thinner when deselected.

Pan now moves completely with the mouse regardless of the zoom level.